Refactor to use a common base class
