+++ django_migration_testcase/django_migrations.py
@@ -1,36 +1,20 @@
 import django
-from django.test import TransactionTestCase
 from django.core.management import call_command
 from django.db import connection
 from django.db.migrations.loader import MigrationLoader
 
+from .base import BaseMigrationTestCase
 
-class MigrationTest(TransactionTestCase):
 
-    __abstract__ = True
-
-    before = None
-    after = None
-    app_name = None
+class MigrationTest(BaseMigrationTestCase):
 
     def setUp(self):
         self.apps_before = {}
         self.apps_after = {}
 
         super(MigrationTest, self).setUp()
-
-        if self.app_name:
-            self.before = [(self.app_name, self.before)]
-            self.after = [(self.app_name, self.after)]
-
         for app_name, version in self.before:
             call_command('migrate', app_name, version,
-                         no_initial_data=True, verbosity=0)
-
-    def tearDown(self):
-        super(MigrationTest, self).tearDown()
-        for app_name, _ in self.after:
-            call_command('migrate', app_name,
                          no_initial_data=True, verbosity=0)
 
     def _get_apps_for_migration(self, app_label, migration_name):
@@ -42,27 +26,20 @@ class MigrationTest(TransactionTestCase):
             state.render()
         return state.apps
 
-    def _get_app_and_model_name(self, model_name):
-        if '.' in model_name:
-            return model_name.split('.', 2)
-        else:
-            return self.app_name, model_name
-
-    def get_model_before(self, model_name):
+    def _get_model(self, model_name, before_or_after, apps_dict):
         app_name, model_name = self._get_app_and_model_name(model_name)
-        version = dict(self.before)[app_name]
-        if app_name not in self.apps_before:
-            self.apps_before[app_name] = self._get_apps_for_migration(app_name, version)
-        return self.apps_before[app_name].get_model(app_name, model_name)
-
-    def get_model_after(self, model_name):
-        app_name, model_name = self._get_app_and_model_name(model_name)
-        version = dict(self.after)[app_name]
-        if app_name not in self.apps_after:
-            self.apps_after[app_name] = self._get_apps_for_migration(app_name, version)
-        return self.apps_after[app_name].get_model(app_name, model_name)
+        version = dict(before_or_after)[app_name]
+        if app_name not in apps_dict:
+            apps_dict[app_name] = self._get_apps_for_migration(app_name, version)
+        return apps_dict[app_name].get_model(app_name, model_name)
 
     def run_migration(self):
         for app_name, version in self.after:
             call_command('migrate', app_name, version,
-                         no_initial_data=True, verbosity=0)
+                         verbosity=0, no_initial_data=True)
+
+    def get_model_before(self, model_name):
+        return self._get_model(model_name, self.before, self.apps_before)
+
+    def get_model_after(self, model_name):
+        return self._get_model(model_name, self.after, self.apps_after)
