+++ django_migration_testcase/south_migrations.py
@@ -1,29 +1,18 @@
-from django.test import TransactionTestCase
 from django.core.management import call_command
 
 from south.migration import Migrations
 
+from .base import BaseMigrationTestCase
 
-class MigrationTest(TransactionTestCase):
+
+class MigrationTest(BaseMigrationTestCase):
     """Test for migrations, reworked from:
     https://micknelson.wordpress.com/2013/03/01/testing-django-migrations/
 
     """
 
-    __abstract__ = True
-
-    before = None
-    after = None
-    app_name = None
-
     def setUp(self):
         super(MigrationTest, self).setUp()
-
-        # if self.app_name is None, then assume self.before is a list
-        # of 2-tuples. This is more explicit (and easier to document).
-        if self.app_name:
-            self.before = [(self.app_name, self.before)]
-            self.after = [(self.app_name, self.after)]
 
         self.before_migrations = []
         for app_name, version in self.before:
@@ -52,22 +41,8 @@ class MigrationTest(TransactionTestCase):
             call_command('migrate', app_name, version,
                          verbosity=0, no_initial_data=True)
 
-    def tearDown(self):
-        # We do need to tidy up and take the database to its final
-        # state so that we don't get errors when the final truncating
-        # happens.
-        for app_name, _ in self.after:
-            call_command('migrate', app_name,
-                         verbosity=0, no_initial_data=True)
-        super(MigrationTest, self).tearDown()
-
     def _get_model(self, model_name, orm_dict):
-        if '.' in model_name:
-            app_name, model_name = model_name.split('.', 2)
-        elif self.app_name:
-            app_name = self.app_name
-        else:
-            raise ValueError('Must specify app name')
+        app_name, model_name = self._get_app_and_model_name(model_name)
         # Because we store all the orms for each migration against
         # their app name, lookup the relevant orm state first.
         orm = orm_dict[app_name]
